feat: Add debug param for skipping to outro
Introduces a 'debug' query parameter to the application's URL. This allows developers or testers to skip directly to the game's outro sequence by appending `?debug=outro` to the URL, bypassing all levels and intros. This is useful for rapid testing of end-game scenarios or for demonstrating the final state of the application.